testing components pt2
