+++ client/src/components/DropDownMenuX.css
@@ -1,0 +1,3 @@
+.prefecture {
+    background-color: #FCD8D4;
+}
