+++ client/src/components/DropdownMenu.css
@@ -2,3 +2,12 @@ select {
     background-color: #FCD8D4;
     border-radius: 5px;
 }
+
+/* select {
+    background-color: #fff;
+    border: 1px solid #ccc;
+  }
+  
+  select option {
+    border: 1px solid #ccc;
+  } */
